FinnhubProducer/requirements.txt
kafka-python-ng==2.2.3
websocket-client==1.4.2
finnhub-python==2.4.15
avro-python3==1.10.2
